unit testing for Description component
