--- client/src/index.js
@@ -1,1 +1,0 @@
-import Description from './components/index.jsx';
