+++ package.json
@@ -9,6 +9,9 @@
   "engines": {
     "node": ">=6.13.0"
   },
+  "jest": {
+    "setupTestFrameworkScriptFile": "<rootDir>/client/setupTests.js"
+   },
   "scripts": {
     "start": "nodemon server/server.js",
     "test": "jest",
@@ -18,28 +21,33 @@
   "dependencies": {
     "axios": "^0.18.0",
     "babel-loader": "^7.1.4",
+    "enzyme-adapter-react-16": "^1.1.1",
     "express": "^4.16.3",
     "faker": "^4.1.0",
+    "jest": "^22.4.3",
     "mongod": "^2.0.0",
     "mongoose": "^5.0.15",
+    "node": "^9.11.0",
     "nodemon": "^1.17.3",
     "path": "^0.12.7",
     "react": "^16.3.1",
     "react-dom": "^16.3.1",
     "react-modal": "^3.3.2",
-    "react-router": "^4.2.0",
-    "react-router-dom": "^4.2.2",
+    "react-test-renderer": "^16.3.2",
     "styled-components": "^3.2.5",
-    "travis": "^0.1.1"
+    "travis": "^0.1.1",
+    "webpack-dev": "^1.1.0"
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
+    "babel-jest": "^22.4.3",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "css-loader": "^0.28.11",
+    "enzyme": "^3.3.0",
     "eslint": "^4.19.1",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
-    "jest": "^22.4.3",
     "node-sass": "^4.8.3",
     "sass-loader": "^6.0.7",
     "style-loader": "^0.20.3",
