+++ webpack.config.js
@@ -19,5 +19,5 @@ module.exports = {
         }
       }
     ]
-  }
+  }, 
 };
